some cleanup and refactor
