+++ packages/event-processor/src/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright 2019, Optimizely
+ * Copyright 2019-2020, Optimizely
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 export * from './events'
 export * from './eventProcessor'
 export * from './eventDispatcher'
 export * from './managed'
 export * from './pendingEventsDispatcher'
 export * from './v1/buildEventV1'
-export * from './v1/v1EventProcessor'
+export { LogTierV1EventProcessor } from './v1/v1EventProcessor'
